[refactor]: clean up
